feat(launcher): uses gtk.spawn to parse desktop files accurately
one strange bug I encountered here was how I was running my statusbar.
since i was running it wit hte LD_PRELOAD variable set, so that
gtk-layershell would actually get picked up, this meant that it was
beraking some other stuff that required gtk.init to be called before the
variable was set. I just remove that form the environment. Thanks to
ULauncher for inspiration on what to do to get this working